refactor(mcp-manager): extract exchanges logic to dedicated class
